musl-libc 2/4, au_decode_mntpnt()
In getmntent(3), glibc decodes the path string, but musl doesn't.
In other words, glibc converts the retreived characters represented in
octal into the corresponding ascii representation.
This commits introduces a new function to decode the characters in octal
representation, in order to make the musl behaviour to be compatible to
glibc.
See also previous commits in this series.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>