musl-libc 3/4, separate libau.so individually
When the system has two versions of libc.so, for instance glibc and
musl-libc, libau.so needs to have two versions. One for glibc and
another is for musl, obviously.
This commit makes libau.so not only to support musl, but also has the
install path individually.
See also previous commits in this series.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>